works out rows and diagonals

+++ magic_square_test.go
@@ -7,9 +7,9 @@ import (
 
 func Test_E2E_Rows(t *testing.T) {
 	ms, _ := CreateSquare()
-	assert.Equal(t, 9, findArraySum(ms.Row1))
-	assert.Equal(t, 9, findArraySum(ms.Row2))
-	assert.Equal(t, 9, findArraySum(ms.Row3))
+	assert.Equal(t, 9.0, findArraySum(ms.Row1))
+	assert.Equal(t, 9.0, findArraySum(ms.Row2))
+	assert.Equal(t, 9.0, findArraySum(ms.Row3))
 }
 
 //func Test_E2E_Columns(t *testing.T) {
@@ -30,24 +30,24 @@ func Test_E2E_Rows(t *testing.T) {
 //	assert.Equal(t, 9, findArraySum(diagonal2))
 //}
 
-func findArraySum(arr []int) int {
-	res := 0
+func findArraySum(arr []float64) float64 {
+	res := 0.0
 	for i := 0; i < len(arr); i++ {
 		res += arr[i]
 	}
 	return res
 }
 
-func createColumns(ms MagicSquare, element int) []int {
-	column := []int{}
+func createColumns(ms MagicSquare, element int) []float64 {
+	column := []float64{}
 	column = append(column, ms.Row1[element])
 	column = append(column, ms.Row2[element])
 	column = append(column, ms.Row3[element])
 	return column
 }
 
-func createDiagonal1(ms MagicSquare) []int {
-	diagonal1 := []int{}
+func createDiagonal1(ms MagicSquare) []float64 {
+	diagonal1 := []float64{}
 	diagonal1 = append(diagonal1, ms.Row1[0])
 	diagonal1 = append(diagonal1, ms.Row2[1])
 	diagonal1 = append(diagonal1, ms.Row3[2])
@@ -55,8 +55,8 @@ func createDiagonal1(ms MagicSquare) []int {
 	return diagonal1
 }
 
-func createDiagonal2(ms MagicSquare) []int {
-	diagonal2 := []int{}
+func createDiagonal2(ms MagicSquare) []float64 {
+	diagonal2 := []float64{}
 	diagonal2 = append(diagonal2, ms.Row1[2])
 	diagonal2 = append(diagonal2, ms.Row2[1])
 	diagonal2 = append(diagonal2, ms.Row3[0])
